msDocumentCompressor/requirements.txt
PyYAML==6.0
py-eureka-client==0.11.7
Flask==2.2.3
docx2txt==0.8
docxcompose==1.4.0
docxtpl==0.16.5
pypdf==3.5.2
Pillow==9.4.0
python-docx==0.8.11
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
ifaddr==0.2.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
pip==22.3.1
py-eureka-client==0.11.7
PyYAML==6.0
setuptools==65.6.3
Werkzeug==2.2.3
wheel==0.38.4
zipp==3.14.0
azure-core==1.26.3
azure-identity==1.12.0
azure-storage-blob==12.15.0
PyPDF2==3.0.1
cffi==1.14.0
